prototypes/python/requirements.txt
jsonschema
pyyaml
opentelemetry-api
opentelemetry-sdk